% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_formatting.R
\name{apply_deleted_row_styling}
\alias{apply_deleted_row_styling}
\title{Apply deleted row styling to DataTable}
\usage{
apply_deleted_row_styling(dt, data)
}
\arguments{
\item{dt}{DataTable object}

\item{data}{Source dataframe (must contain deleted_by_human column)}
}
\value{
DataTable with deleted row styling applied
}
\description{
Apply deleted row styling to DataTable
}
