% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_formatting.R
\name{apply_edited_styling}
\alias{apply_edited_styling}
\title{Apply edited cell styling to DataTable}
\usage{
apply_edited_styling(dt, data, edited_cells)
}
\arguments{
\item{dt}{DataTable object}

\item{data}{Source dataframe}

\item{edited_cells}{Dataframe with interaction_id and column_name}
}
\value{
DataTable with edited cell styling
}
\description{
Apply edited cell styling to DataTable
}
