% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_formatting.R
\name{apply_restored_row_styling}
\alias{apply_restored_row_styling}
\title{Apply restored interaction styling to DataTable}
\usage{
apply_restored_row_styling(dt, data, restored_ids)
}
\arguments{
\item{dt}{DataTable object}

\item{data}{Source dataframe}

\item{restored_ids}{Vector of interaction IDs that are restored}
}
\value{
DataTable with restored row styling applied
}
\description{
Apply restored interaction styling to DataTable
}
