% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_highlighting.R
\name{calculate_sentence_similarities}
\alias{calculate_sentence_similarities}
\title{Calculate similarities between target sentence and candidate sentences}
\usage{
calculate_sentence_similarities(
  candidate_sentences,
  clean_target,
  source_words
)
}
\arguments{
\item{candidate_sentences}{Vector of candidate sentences}

\item{clean_target}{Target sentence to match}

\item{source_words}{Words from the target sentence}
}
\value{
Vector of similarity scores
}
\description{
Calculate similarities between target sentence and candidate sentences
}
\keyword{internal}
