% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.R
\name{cleanup_temp_files}
\alias{cleanup_temp_files}
\title{Clean up old temporary files}
\usage{
cleanup_temp_files(current_session_token = NULL)
}
\arguments{
\item{current_session_token}{Current session token to preserve}
}
\description{
Clean up old temporary files
}
