% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_conversion.R
\name{convert_edit_value}
\alias{convert_edit_value}
\title{Convert edited cell value to the appropriate type}
\usage{
convert_edit_value(value, target_type)
}
\arguments{
\item{value}{The new value as a string from the cell edit}

\item{target_type}{The class of the target column}
}
\value{
The converted value in the appropriate type
}
\description{
Convert edited cell value to the appropriate type
}
