% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_combine.R
\name{.fetch_by_doc_id}
\alias{.fetch_by_doc_id}
\title{Fetch rows from a table filtered by a vector of document_ids}
\usage{
.fetch_by_doc_id(con, tbl, doc_ids)
}
\description{
Fetch rows from a table filtered by a vector of document_ids
}
\keyword{internal}
