% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_combine.R
\name{.fetch_by_interaction_id}
\alias{.fetch_by_interaction_id}
\title{Fetch rows from a table filtered by a vector of interaction ids}
\usage{
.fetch_by_interaction_id(con, tbl, interaction_ids)
}
\description{
Fetch rows from a table filtered by a vector of interaction ids
}
\keyword{internal}
