% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_highlighting.R
\name{extract_sentences_from_markdown}
\alias{extract_sentences_from_markdown}
\title{Extract sentences from HTML or markdown text}
\usage{
extract_sentences_from_markdown(text)
}
\arguments{
\item{text}{HTML or markdown text to process}
}
\value{
Vector of sentences
}
\description{
Extract sentences from HTML or markdown text
}
