% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_formatting.R
\name{get_all_edited_cells}
\alias{get_all_edited_cells}
\title{Get all edited cells for styling}
\usage{
get_all_edited_cells(document_id, pending_edits)
}
\arguments{
\item{document_id}{Document ID to get edits for}

\item{pending_edits}{List of pending edits not yet saved}
}
\value{
Dataframe with interaction_id and column_name columns
}
\description{
Get all edited cells for styling
}
