% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_formatting.R
\name{get_restored_interaction_ids}
\alias{get_restored_interaction_ids}
\title{Get restored interaction IDs for styling}
\usage{
get_restored_interaction_ids(document_id)
}
\arguments{
\item{document_id}{Document ID to check}
}
\value{
Vector of interaction IDs that are restored
}
\description{
Get restored interaction IDs for styling
}
