% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_highlighting.R
\name{inject_row_spans}
\alias{inject_row_spans}
\title{Wrap cells of one table row with evidence spans (internal)}
\usage{
inject_row_spans(html, row, cells, span_open)
}
\arguments{
\item{html}{HTML string}

\item{row}{Index of the \code{<tr>} in \code{html}}

\item{cells}{Indices of the cells in that row to wrap}

\item{span_open}{Opening span tag}
}
\value{
Modified HTML
}
\description{
Wrap cells of one table row with evidence spans (internal)
}
\keyword{internal}
