% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_trail.R
\name{process_delete_with_renumbering}
\alias{process_delete_with_renumbering}
\title{Process delete with renumbering}
\usage{
process_delete_with_renumbering(interaction_id, document_id, db_conn = NULL)
}
\arguments{
\item{interaction_id}{ID of interaction to delete}

\item{document_id}{Document ID}

\item{db_conn}{Database connection path}
}
\value{
List with success status and renumber results
}
\description{
Process delete with renumbering
}
